services/model_service/pyproject.toml
[project]
name = "model-service"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn",
    "httpx",
    "huggingface-hub",
    "pydantic-settings>=2.13.1",
    "common",
]

[dependency-groups]
heavy = [
    "torch",
    "sentence-transformers",
]

[tool.uv.sources]
common = { workspace = true }